+++ domain/src/main/kotlin/br/com/amz/replay/offer/model/Offer.kt
@@ -1,16 +1,17 @@
 package br.com.amz.replay.offer.model
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class Offer(
     val id: UUID,
-    val monthlyPaymentAmount: Double,
+    val monthlyPaymentAmount: BigDecimal,
     val annualPercentageRate: Double,
     val paymentAmount: Int
 )
 
 data class ProposalOffer(
     val annualPercentageRate: Double,
-    val monthlyPayment: Double,
+    val monthlyPayment: BigDecimal,
     val timeRemaining: Int,
 )
