+++ input/rest/src/main/kotlin/br/com/amz/replay/offer/dto/OfferInputDTO.kt
@@ -2,11 +2,12 @@ package br.com.amz.replay.offer.dto
 
 import br.com.amz.replay.offer.model.Offer
 import br.com.amz.replay.offer.model.ProposalOffer
+import java.math.BigDecimal
 import java.util.UUID
 
 data class OfferDTO (
     val id: UUID,
-    val monthlyPaymentAmount: Double,
+    val monthlyPaymentAmount: BigDecimal,
     val annualPercentageRate: Double,
     val paymentAmount: Int,
 )
@@ -20,7 +21,7 @@ fun Offer.toDTO() = OfferDTO(
 
 data class ProposalOfferDTO (
     val annualPercentageRate: Double,
-    val monthlyPayment: Double,
+    val monthlyPayment: BigDecimal,
     val timeRemaining: Int
 )
